% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bsModal.R
\name{bsModal2}
\alias{bsModal2}
\title{bsModal2}
\usage{
bsModal2(id, title, trigger, ..., class = NULL, size)
}
\arguments{
\item{id}{A unique identifier for the modal window}

\item{title}{The title to appear at the top of the modal}

\item{trigger}{The id of a button or link that will open the modal.}

\item{class}{Class to customize look via CSS. \code{NULL} for the default \code{shinyBS} look, or a named vector or list with one or more classes applied to fade, content, header, close (the ``x'' at upper right), title, body, footer, and btn elements.}

\item{size}{Optional What size should the modal be? (small or large)}

\item{\dots}{UI elements to include within the modal}
}
\description{
\code{bsModal2} is used within the UI to create a modal window.
}
\details{
See \code{\link[shinyBS]{Modals}} for more information about how to use the parent function, \code{bsModal}, with the rest of the Modals family. The \code{class} argument was added by Jon Katz.
}
\note{
Run \code{bsExample("Modals")} for an example of \code{bsModal} functionality.
}
\author{
Eric Bailey
}
\seealso{
\href{http://getbootstrap.com/}{Twitter Bootstrap 3}
}
\keyword{manip}

