% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/carouselUI.R
\name{carouselUI}
\alias{carouselUI}
\title{carouselUI}
\usage{
carouselUI(slides, data.interval = 10000, data.ride = "carousel")
}
\arguments{
\item{slides}{A list of HTML tag elements, e.g. \code{list(shiny::div(), shiny::div())}.}

\item{data.interval}{Milliseconds to pause before auto advancing (numeric).}

\item{data.ride}{Either \code{carousel} or \code{NA}; unclear if this works.}
}
\value{
HTML code with class \code{shiny.tag} that can be passed to \code{uiOutput}.
}
\description{
Make a dynamic carousel for use with shiny::renderUI() and shiny::uiOutput().
}
\note{
Adapted by JEK from \url{https://github.com/dcurrier/carouselPanel/blob/master/carouselPanel.R}
}
\examples{
c.slides <- lapply(1:5, function(x) div(h3(style='text-align:center;',paste('Slide', x))))
carouselUI(slides=c.slides, data.interval=5000, data.ride=FALSE)

\dontrun{
/*     place in www/css/styles.css:    */
    .carousel-indicators li {
        background-color: #DDD;
        background-color: rgba(70,70,70,.25);
    }
    .carousel-indicators .active {
        background-color: #999;
    }
    .carousel-indicators {
        bottom: 15px !important;
        top: auto;
        list-style: none outside none;
        margin: 0;
        position: absolute;
        right: 70px;
        z-index: 5;
    }
    .carousel-control {
        opacity: 0.2;
        border: none;
        top: auto;
    }
    .item {
        padding: 0px 17px;
    }
}
}
\author{
Jon Katz
}
\keyword{manip}

