% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inlineInputs.R
\name{inumericInput}
\alias{inumericInput}
\title{Create a numeric input with a label either left, above, or both left and above the input. With no label it becomes an 'inline' input suitable for embedding in a table.}
\usage{
inumericInput(inputId, label, label2 = NULL, value, min = NA, max = NA,
  step = NA, label.style = "float:left;padding:0.5em 1em 0.5em 0em;",
  input.style = "width:75\%;border-radius:4px;",
  container.style = "margin-bottom:0px;", container = list(outer =
  shiny::div, inner = shiny::span))
}
\arguments{
\item{inputId}{The DOM ID.}

\item{label}{The label to display to the left of the input.}

\item{label2}{The optional label to display to the above of the input.}

\item{value}{The initial value for the input.}

\item{min}{The minimum acceptable value for the input.}

\item{max}{The maximum acceptable value for the input.}

\item{step}{The step increment value for the input.}

\item{label.style}{Inline css to apply to the left-of-input label.}

\item{input.style}{Inline css to apply to the main input tag.}

\item{container.style}{Inline css to apply to the outer container.}

\item{container}{A named list of containers to apply. May length one or two, with names in c('outer', 'inner'); if length one the value will be applied to the outer container, and the inner will be a span by default. Should be either a block container 'div' or an inline container 'span'.}
}
\description{
Create a numeric input with a label either left, above, or both left and above the input. With no label it becomes an 'inline' input suitable for embedding in a table.
}
\examples{
inumericInput(inputId='test', label='test label', value=1)
}
\author{
Jon Katz
}
\keyword{misc}

