% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inlineInputs.R
\name{iradioButtons}
\alias{iradioButtons}
\title{Create a radio button input with a label to the left of the input. With no label it becomes an 'inline' input suitable for embedding in a table.}
\usage{
iradioButtons(inputId, label, choices, selected = NULL, inline = FALSE,
  width = NULL, padding = "0.5em 1em 0.5em 0em", thin = FALSE,
  container = shiny::div)
}
\arguments{
\item{inputId}{The DOM ID.}

\item{label}{The label to display to the left of the input.}

\item{choices}{Named or unnamed character vector of selectable choices.}

\item{selected}{The initial value for the input.}

\item{inline}{Place options inline (TRUE) or stacked (FALSE).}

\item{container}{A named list of containers to apply. May length one or two, with names in c('outer', 'inner'); if length one the value will be applied to the outer container, and the inner will be a span by default. Should be either a block container 'div' or an inline container 'span'.}

\item{label.style}{Inline css to apply to the left-of-input label.}

\item{input.style}{Inline css to apply to the main input tag.}

\item{container.style}{Inline css to apply to the outer container.}
}
\description{
Create a radio button input with a label to the left of the input. With no label it becomes an 'inline' input suitable for embedding in a table.
}
\examples{
iradioButtons(inputId='test', label='test label', value=1)
}
\author{
Jon Katz
}
\keyword{misc}

