% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/popoverInput.R
\name{popoverInput}
\alias{popoverInput}
\title{popoverInput}
\usage{
popoverInput(fnType, id, label, ..., helpText, divStyle = NULL,
  helpStyle = "padding-top:1.5em", popTrigger = "hover", colWidths = c(2,
  10))
}
\arguments{
\item{fnType}{function or function name (character) of desired input, e.g. textInput or 'textInput'}

\item{id}{InputID for shiny element (character).}

\item{label}{Input label (character).}

\item{helpText}{Text to place in popover (character).}

\item{divStyle}{Inline CSS for main div returned (character or NULL).}

\item{helpStyle}{Inline CSS for the help icon (e.g. padding to align it vertically with input; character or NULL).}

\item{popTrigger}{Trigger for popover. In c('hover','focus','click','manual').}

\item{colWidths}{Length 2 numeric vector specifying width of help button column and width of input function column. Should sum to 12.}

\item{\dots}{Additional named arguments to input function.}
}
\value{
HTML code.
}
\description{
Place a help icon with popover text next to a shiny input function
}
\note{
Requires Eric Bailey's shinyBS package (available from CRAN).
}
\examples{
popoverInput(
       checkboxGroupInput,
       'lyrOpts',
       'Options:',
       choices=c(
           'Paste URL'='url',
           'Clip to Layer'='clip',
           'Projection Ref'='projection',
           'No Data'='empty'
           ),
       selected=c('clip','projection'),
       inline=TRUE,
       helpText="To download a layer from a website check \\"Paste URL\\". You may also clip or project all rasters to match this one. If the layer contains only boundary or projection information check \\"No Data\\".",
       divStyle='padding-top:1em;padding-bottom:1em;',
       helpStyle='padding-top:0.5em;'
   )
}
\author{
Jon Katz
}
\keyword{manip}

