% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/popoverInput.R
\name{popoverInput2}
\alias{popoverInput2}
\title{popoverInput2}
\usage{
popoverInput2(fnType, ..., helpText, fa.icon = "fa-question-circle",
  fa.color = "#337ab7", padding.top = "0.3em", popTrigger = "hover click",
  container = shiny::div, cStyle = "clear:both;")
}
\arguments{
\item{fnType}{function or function name (character) of desired input, e.g. textInput or 'textInput'}

\item{helpText}{Text to place in popover (character). Omit to align an helpless input with a helpful input.}

\item{fa.icon}{Icon name from font awesome, default is "fa-question-circle".}

\item{fa.color}{Color of icon; defaults to bootstrap link color.}

\item{padding.top}{Padding to align the help icon vertically with input label.}

\item{popTrigger}{Trigger for popover. In c('hover','focus','click','manual').}

\item{container}{HTML container to hold the input, either \code{div} or \code{span}}

\item{cStyle}{Inline CSS for main container returned (character or NULL).}

\item{id}{InputID for shiny element (character).}

\item{label}{Input label (character).}

\item{\dots}{Additional named arguments to input function.}
}
\value{
HTML code.
}
\description{
Place a help icon with popover text next to a shiny input function. Both the help icon and the input element float left.
}
\note{
Requires Eric Bailey's shinyBS package (available from CRAN).
}
\examples{
popoverInput2(
       checkboxGroupInput,
       'lyrOpts',
       'Options:',
       choices=c(
           'Paste URL'='url',
           'Clip to Layer'='clip',
           'Projection Ref'='projection',
           'No Data'='empty'
           ),
       selected=c('clip','projection'),
       inline=TRUE,
       helpText="To download a layer from a website check \\"Paste URL\\". You may also clip or project all rasters to match this one. If the layer contains only boundary or projection information check \\"No Data\\".",
       padding.top="1em"
   )
}
\author{
Jon Katz
}
\keyword{manip}

