% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rName.R
\name{rName}
\alias{rName}
\title{rName}
\usage{
rName(char = 16, ext = NULL)
}
\arguments{
\item{char}{The number of characters, not including the optional extension (numeric).}

\item{ext}{The optional extension, starting with a dot e.g. '.txt' (character).}
}
\description{
Generate names composed of random characters and numbers. There is no guarantee names won't collide, but they are guaranteed to start with a letter.
}
\note{
Kind of similar to \code{basename(tempfile("",""))}, but less rigorous and not hex based.
}
\examples{
rName()
}
\keyword{manip}

