\name{shinyAddOns-package}
\alias{shinyAddOns-package}
\alias{shinyAddOns}
\docType{package}
\title{
\packageTitle{shinyAddOns}
}
\description{
\packageDescription{shinyAddOns}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{shinyAddOns}
\packageIndices{shinyAddOns}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{shinyAddOns}

Maintainer: \packageMaintainer{shinyAddOns}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
