% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyAddOns.R
\docType{package}
\name{shinyAddOns}
\alias{shinyAddOns}
\alias{shinyAddOns-package}
\title{A Mish-Mash Of Functions Useful In Shiny Apps}
\description{
These functions aren't united by anything they share other than
that they fill a need in my personal apps.
}

