% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uiSetSnapshot.R
\name{uiSetSnapshot}
\alias{uiSetSnapshot}
\title{uiSetSnapshot}
\usage{
uiSetSnapshot(snapshot, sess)
}
\arguments{
\item{snapshot}{A Named List Of Input Types And Values.}

\item{sess}{The shiny \code{Session} object necessary to update values.}
}
\value{
Returns a list of update results.
}
\description{
Bulk Set UI Element Values Given A Named List Of Input Types And Values
}
\examples{
\dontrun{
# Must be called within a shiny app
library(shiny)
ui <- fluidPage(
    textInput("main", "Plot title:", value="First Plot"),
    numericInput("randn", 'Number of Random Values:', value=10, min=1, step=1),
    column(1, actionButton('submit', 'Submit')),
    column(1, actionButton('reset', 'Previous Settings')),
    div(style='clear:both;',
        column(9, plotOutput('outplot')),
        column(3, htmlOutput('snapshot', container=tags$pre, class='shiny-text-output'))
    )
)

server <- function(input, output, session) {
    snapshot <- list()

    observe({
        input$reset
        if(length(snapshot) > 0) {
            s.el <- length(snapshot)
            if(length(snapshot) > 1) snapshot <- snapshot[(s.el - 1):s.el]
            uiSetSnapshot(snapshot[[1]], session)
        }
        NULL
    })

    output$outplot <- renderPlot({
        input$submit
        isolate({
            input.funs <- c(main='textInput', randn='numericInput')
            snapshot <<- c(snapshot, list(uiTakeSnapshot(input.funs, environment())))
            values <- rnorm(input$randn)
            plot(values, main=input$main)
        })
    })

    output$snapshot <- renderPrint({
        input$submit
        input$reset
        if(length(snapshot)>1) {
            print(snapshot[[length(snapshot)-1]])
        } else print(snapshot[[length(snapshot)]])
    })
}
shinyApp(ui, server)
}
}
\author{
Jon Katz
}
\keyword{misc}

