% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uiTakeSnapshot.R
\name{uiTakeSnapshot}
\alias{uiTakeSnapshot}
\title{uiTakeSnapshot}
\usage{
uiTakeSnapshot(inputs, env)
}
\arguments{
\item{inputs}{A named list or vector of input types (character), named with the inputId.}

\item{env}{Environment in which to find the shiny \code{input} list.}
}
\value{
A list of lists, with variable content according to the following structure:
\describe{
#'     \item{inputId}{\describe{
        \item{type}{Input function, as a character string.}
        \item{value}{Value of the UI input at time of function call.}
    }}
}
}
\description{
Bulk Grab UI Element Values Given A List Of Names and Input Types
}
\details{
The output is designed to easily re-create the current state of the UI, setting values with the appropriate \code{update__Input} function. Note that it is not possible to set a \code{fileInput} value, but it is possible to store the name of the uploaded file.
}
\examples{
\dontrun{
# Must be called within a shiny app
library(shiny)
ui <- fluidPage(
    textInput("main", "Plot title:", value="First Plot"),
    numericInput("randn", 'Number of Random Values:', value=10, min=1, step=1),
    column(1, actionButton('submit', 'Submit')),
    column(1, actionButton('reset', 'Previous Settings')),
    div(style='clear:both;',
        column(9, plotOutput('outplot')),
        column(3, htmlOutput('snapshot', container=tags$pre, class='shiny-text-output'))
    )
)

server <- function(input, output, session) {
    snapshot <- list()

    observe({
        input$reset
        if(length(snapshot) > 0) {
            s.el <- length(snapshot)
            if(length(snapshot) > 1) snapshot <- snapshot[(s.el - 1):s.el]
            uiSetSnapshot(snapshot[[1]], session)
        }
        NULL
    })

    output$outplot <- renderPlot({
        input$submit
        isolate({
            input.funs <- c(main='textInput', randn='numericInput')
            snapshot <<- c(snapshot, list(uiTakeSnapshot(input.funs, environment())))
            values <- rnorm(input$randn)
            plot(values, main=input$main)
        })
    })

    output$snapshot <- renderPrint({
        input$submit
        input$reset
        if(length(snapshot)>1) {
            print(snapshot[[length(snapshot)-1]])
        } else print(snapshot[[length(snapshot)]])
    })
}
shinyApp(ui, server)
}
}
\author{
Jon Katz
}
\keyword{misc}

