% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/updateTextareaInput.R
\name{updateTextareaInput}
\alias{updateTextareaInput}
\title{updateTextareaInput}
\usage{
updateTextareaInput(session, inputId, label = NULL, value = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object}

\item{value}{The value to set for the input object.}
}
\description{
Set the value of a textarea on a shiny application UI.
}
\details{
Although a \code{textareaInput} doesn't exist in shiny it is possible to create one with HTML and update it using the shiny messaging system. This function appropriates the code from the shiny updateTextInput.
}
\examples{
# textarea can be created in UI with HTML:
    HTML({"
        <label for='inText'>Old label</label>
        <textarea id='inText', class='textarea', placeholder= ''></textarea>"
    })

\dontrun{
# From ?updateTextInput
shinyServer(function(input, output, session) {

  observe({
    # We'll use the input$controller variable multiple times, so save it as x
    # for convenience.
    x <- input$controller

    # This will change the value of input$inText, based on x
    updateTextareaInput(session, "inText", value = paste("New text", x))

    # Can also set the label, this time for input$inText2
    updateTextareaInput(session, "inText2",
      label = paste("New label", x),
      value = paste("New text", x))
  })
})

}
}

