% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/with_callAttr.R
\name{with_callAttr}
\alias{with_callAttr}
\title{with_callAttr}
\usage{
with_callAttr(expr, f)
}
\arguments{
\item{expr}{An expression to evaluate and capture as a call.}

\item{f}{The frame number, specified as an integer. Passed directly to \code{sys.call}.}
}
\value{
The expected output of \code{expr} with an attribute named 'call', accessible with \code{attr(expr, 'call')}.
}
\description{
Capture the System Call For Functions That Don't Already Do So
}
\examples{
test.expr <- with_callAttr(cbind(x=1:4, y=rnorm(4)), 0)
attr(test.expr,'call')
}
\author{
Jon Katz
}
\keyword{misc}

