% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uiSetSnapshot.R
\name{uiSetSnapshot}
\alias{uiSetSnapshot}
\title{uiSetSnapshot}
\usage{
uiSetSnapshot(snapshot, sess)
}
\arguments{
\item{snapshot}{A Named List Of Input Types And Values.}

\item{sess}{The shiny \code{Session} object necessary to update values.}
}
\value{
Returns a list of update results.
}
\description{
Bulk Set UI Element Values Given A Named List Of Input Types And Values
}
\examples{
\dontrun{
# Must be called within a shiny app
ui <- fluidPage(
    textInput("main", "Plot title:"),
    numericInput("randn", 'Number of Random Values:')
    actionButton('reset', 'Previous Plot')
    plotOutput('outplot')
)

server <- function(input, output, session) {
    snapshot <- NA

    observe({
        lastPlot()
    })

    lastPlot <- reactive({
        input$reset
        if(!is.na(snapshot)) uiSetSnapshot(snapshot, session)
        NULL
    })

    output$outplot <- renderPlot({
        input.funs <- c(main='textInput', randn='numericInput')
        snapshot <<- uiTakeSnapshot(input.funs, environment())
        values <- rnorm(input$randn)
        plot(values, main=input$main)
    })
}
}
}
\author{
Jon Katz
}
\keyword{misc}

